producer/requirements.txt
websocket-client==0.56.0
kafka-python==1.4.6
aiohttp==3.6.1